chore: Fix missing index.html file in the built output
This was removed in #131 in 2019, presumably to fix the sphinx warning
`/workspace/index.rst: WARNING: document isn't included in any toctree`
(the build results from 2019 are no longer available).

We merged that PR in 2024 as it looked reasonable to people unfamiliar
with sphinx.

However it was the wrong fix, as it prevents sphinx generating an
index.html at all. The actual content is all in the theme template
for the index page, but the empty file is required to have that built.

The reinstated file with the `:orphan:` tag silences the warning and
makes the build pass.